requirements.txt
cryptography==2.8
python-dotenv==0.20.0
requests==2.22.0
